"Refactor user data handling and registration process
This commit changes how user data is managed, moving towards using an object-oriented user model instead of manipulating individual properties. This means that instead of passing a GUID and password separately to the registerUser function, a User object is now passed. In addition to supporting better data encapsulation and reducing the amount of code, it also makes the function signature more descriptive and intuitive. Moreover, this commit cleans up the validation process in the user_information, login, and registration_form views to remove redundancy and improve error message accuracy, to provide a better user experience."