"Added register link to login page
A registration link has been added to the login page in order to improve user navigation. Users without accounts can now easily access the registration form directly from the login page. This change is made to enhance the user experience and reduce the time taken for a new user to register."