"Implement user data update functionality
Added the ability for users to update their data. The updateUser function was modified in the backend/database.php to execute an SQL UPDATE command on the 'users' table. The frontend/views/user_information.php was also updated to utilize the new function signature and to redirect to a new page, userDataUpdateSuccessful.php, upon successful update. The new HTML page provides a message confirming the success of the user data update."