Refactor database schema for efficiency and simplicity
In this commit, the database schema has been significantly refactored to enhance simplicity and efficiency. The user_id column was replaced with guid as the PRIMARY KEY in the users table, reducing redundancy. References to user_id were replaced with guid in all other tables as well. Several columns like 'name', 'payment_card_number', and 'is_admin' were removed from the users table due to being unused in the current use-case of the database, simplifying the schema. The transactions_history table was also removed due to overlapping functionality with the user_bought_items table. We restructured foreign keys in item_categories for more optimized queries and replaced category_id with category_name for better readability.