Refactor Item Class and modify database related files
The 'item' class was refactored to 'Item' class to follow the standard class naming conventions in PHP. All includes of "item.php" were updated to "Item.php". The 'Item' class was enhanced with an 'item_id' property for better data manipulation, which also replaced the usage of itemId variable in database.php. The 'item_id' is now directly accessed from the 'Item' object when querying the database, which makes the code cleaner and maintainable.