Replace user GUID with auto-incrementing user ID
Implemented a significant overhaul in the way users are referenced across the system by replacing the GUID based identification with an auto-incrementing user Id. This change enhances stability and accessibility as user identities are less prone to change. Adopting this approach also simplifies the process of user-related operations such as querying and fetching items, registration and linking items to users. The update extends over multiple files including backend and frontend logic, database procedures and data structures. The user's GUID can still change over time, but the new identifier remains constant for the user's lifetime, providing a more reliable reference to each user in the system.