Refactor picture input validation for dynamic field names
The 'validatePictureInput' function in 'inputFieldsValidator.php' has been refactored to validate picture upload fields with dynamic names instead of a static 'profile_pic' field name. With this change, the function is now more flexible and can be used across different picture upload scenarios. The 'profile_pic' field name was replaced by '$pictureFieldName' variable passed as an argument to the function. This won't affect existing functionality but enhances the reusability of the function.